Changed handling of pressure dependent reactions.
Pressure-dependent rxns, when split into irreversible rxns, now correctly use 'high' or 'low' based on whether bimolecular or unimolecular. This includes correcting based on the forward direction of the original reaction (as displayed). Reverse coefficients are calculated for the high or low-pressure limit coefficients.

This involved changing calc_rev_Arrhenius to accept any three coefficients (in a list) and return three reverse coefficients (in a list), to handle accepting low/high-pressure limit coefficients.